task(SDK-3952) - Fixes ClassCastException in deprecated feature ProductConfig

+++ android/src/main/java/com/clevertap/clevertap_plugin/CleverTapPlugin.java
@@ -1817,7 +1817,7 @@ public class CleverTapPlugin implements ActivityAware,
     }
 
     private void setMinimumFetchIntervalInSeconds(MethodCall call, Result result) {
-        long interval = call.argument("interval");
+        int interval = call.argument("interval");
         if (isCleverTapNotNull(cleverTapAPI)) {
             cleverTapAPI.productConfig().setMinimumFetchIntervalInSeconds(interval);
             result.success(null);
